forgot size_t things

+++ src/cute_file_system.h
@@ -42,7 +42,7 @@ struct stat_t
 {
 	file_type_t type;
 	int is_read_only;
-	uint64_t size;
+	size_t size;
 	uint64_t last_modified_time;
 	uint64_t created_time;
 	uint64_t last_accessed_time;
@@ -63,15 +63,15 @@ extern CUTE_API error_t CUTE_CALL file_system_create_dir(const char* virtual_pat
 extern CUTE_API const char** CUTE_CALL file_system_enumerate_directory(const char* virtual_path);
 extern CUTE_API void CUTE_CALL file_system_free_enumerated_directory(const char** directory_list);
 extern CUTE_API error_t CUTE_CALL file_system_file_exists(const char* virtual_path);
-extern CUTE_API uint64_t CUTE_CALL file_system_read(file_t* file, void* buffer, uint64_t bytes);
-extern CUTE_API uint64_t CUTE_CALL file_system_write(file_t* file, const void* buffer, uint64_t bytes);
+extern CUTE_API size_t CUTE_CALL file_system_read(file_t* file, void* buffer, size_t bytes);
+extern CUTE_API size_t CUTE_CALL file_system_write(file_t* file, const void* buffer, size_t bytes);
 extern CUTE_API error_t CUTE_CALL file_system_eof(file_t* file);
-extern CUTE_API uint64_t CUTE_CALL file_system_tell(file_t* file);
-extern CUTE_API error_t CUTE_CALL file_system_seek(file_t* file, uint64_t position);
-extern CUTE_API uint64_t CUTE_CALL file_system_size(file_t* file);
+extern CUTE_API size_t CUTE_CALL file_system_tell(file_t* file);
+extern CUTE_API error_t CUTE_CALL file_system_seek(file_t* file, size_t position);
+extern CUTE_API size_t CUTE_CALL file_system_size(file_t* file);
 extern CUTE_API error_t CUTE_CALL file_system_flush(file_t* file);
-extern CUTE_API error_t CUTE_CALL file_system_read_entire_file_to_memory(const char* virtual_path, void** data_ptr, uint64_t* size = NULL, void* user_allocator_context = NULL);
-extern CUTE_API error_t CUTE_CALL file_system_write_entire_buffer_to_file(const char* virtual_path, const void* data, uint64_t size);
+extern CUTE_API error_t CUTE_CALL file_system_read_entire_file_to_memory(const char* virtual_path, void** data_ptr, size_t* size = NULL, void* user_allocator_context = NULL);
+extern CUTE_API error_t CUTE_CALL file_system_write_entire_buffer_to_file(const char* virtual_path, const void* data, size_t size);
 
 }
